superconsole

Module components

Getting started truly from scratch is excessive for every user of superconsole. A small set of starter components are provided, exposed below.

Re-exports

pub use alignment::Aligned;

pub use bordering::Bordered;

pub use padding::Padded;

pub use splitting::Split;

Modules

alignment

Oftentimes it is useful to align text inside the bounding box. This component offers a composiion based approach to do so. One may align text left, center, or right, and top, middle, or bottom. Additionally, horizontally left aligned text may be optionally justified.

bordering

padding

splitting

Splitting is one of the most primitive building blocks of a fully featured UI. This module contains components and enums that allow for splits along either dimension.

Structs

Blank

The Blank component is a dead-end component that emits nothing. It can be used for testing purposes or to make a portion of a dividing component empty. By default, the SuperConsole is created with a Blank component.

Bounded

Component that ensures its child component has at most max_size render space.

DrawHorizontal

Draw components horizontally, each next component is stacked on the right.

DrawVertical

Draw components vertically, one after the other.

Enums

DrawMode

Used to mark whether a draw is final.

Traits

Component

Components are pluggable drawers that output lines of formatted text. They are composable (eventually) and re-render in place at each render.