superconsole::components

Trait Component

pub trait Component {
    // Required method
    fn draw_unchecked(
        &self,
        dimensions: Dimensions,
        mode: DrawMode,
    ) -> Result<Lines>;

    // Provided method
    fn draw(&self, dimensions: Dimensions, mode: DrawMode) -> Result<Lines> { ... }
}

Components are pluggable drawers that output lines of formatted text. They are composable (eventually) and re-render in place at each render.

Required Methods

fn draw_unchecked( &self, dimensions: Dimensions, mode: DrawMode, ) -> Result<Lines>

This method is to be implemented for components to provide the draw method.

Provided Methods

fn draw(&self, dimensions: Dimensions, mode: DrawMode) -> Result<Lines>

Interprets the current caller state to create its drawing. Dimensions refers to the maximum (width, height) this component may use. The mode refers to if this is the final time the component will be drawn. If a child component is too large to fit in the dimensions, it is truncated.