superconsole::content

Struct Span

#[non_exhaustive]
pub struct Span {
    pub style: ContentStyle,
    pub hyperlink: Option<Hyperlink>,
    /* private fields */
}

A Span is a segment of text that may or may not have style applied to it. One may think of this as a unit of text. It must be valid to be constructed.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

style: ContentStylehyperlink: Option<Hyperlink>
Implementations

impl Span

pub fn dash() -> Span

pub fn valid(stringlike: &str) -> bool

Spans are only valid if they do not contain any non whitespace characters other than a space. This is because the superconsole expects content to be mono-spaced and newlines, tabs, etc., violate this.

pub fn sanitized<S: Display>(string: S) -> Self

pub fn content(&self) -> &str

pub fn padding(amount: usize) -> Self

Create an unstyled span with the specified amount of whitespace padding.

pub fn is_mergeable_with(&self, other: &Span) -> bool

Determine if this span is mergeable with another span, i.e. if they are equal except for content.

pub fn new_unstyled<S: Display>(stringlike: S) -> Result<Self, SpanError>

Attempt to create a new, unstyled span equivalent to the underlying stringlike. This will fail if the input string is not valid.

pub fn new_unstyled_lossy<S: Display>(stringlike: S) -> Self

pub fn new_styled(content: StyledContent<String>) -> Result<Self, SpanError>

Equivalent to new_unstyled, except with styling.

pub fn new_styled_lossy(span: StyledContent<String>) -> Self

Equivalent to Self::new_styled except it will sanitize the content.

pub fn new_styled_raw(span: StyledContent<String>) -> Self

Create a Span that allows all whitespace characters, specially for emit/emit_aux scenarios.

Note: This method bypass whitespace validation, and should not be used for layout components. Itโs specially designed for emit/emit_aux where content is displayed above the canvas and doesnโt require percise characters alignment.

pub fn new_colored(text: &str, color: Color) -> Result<Self, SpanError>

pub fn new_colored_lossy(text: &str, color: Color) -> Self

pub fn with_hyperlink(self, hyperlink: Option<Hyperlink>) -> Self

pub fn len(&self) -> usize

Returns the number of graphemes in the span.

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> impl Iterator<Item = Span> + '_

Iterates over each Grapheme in the Span. Applies the stylization of the Span to each Grapheme. Because a Grapheme is represented as another string, the sub-Span is represented as a Span. This panics if it encounters unicode that it doesnโt know how to deal with.

pub fn fmt_for_test(&self) -> impl Display + '_