superconsole::output

Trait SuperConsoleOutput

pub trait SuperConsoleOutput:
    Send
    + Sync
    + 'static {
    // Required methods
    fn should_render(&mut self) -> bool;
    fn output(&mut self, buffer: Vec<u8>) -> Result<()>;
    fn finalize(self: Box<Self>) -> Result<()>;
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;

    // Provided methods
    fn output_to(&mut self, buffer: Vec<u8>, target: OutputTarget) -> Result<()> { ... }
    fn aux_stream_is_tty(&self) -> bool { ... }
    fn terminal_size(&self) -> Result<Dimensions> { ... }
}

Required Methods

fn should_render(&mut self) -> bool

Called before rendering will occur. This has a chance to prevent rendering by returning false.

fn output(&mut self, buffer: Vec<u8>) -> Result<()>

Called to produce output. This may be called without should_render if we are finalizing or clearing. This should flush if possible.

fn finalize(self: Box<Self>) -> Result<()>

Called when the console has finalized. This must block if necessary. No further output will be emitted.

fn as_any(&self) -> &dyn Any

Get this Output as an Any. This is used for testing.

fn as_any_mut(&mut self) -> &mut dyn Any

Get this Output as a mutable Any. This is used for testing.

Provided Methods

fn output_to(&mut self, buffer: Vec<u8>, target: OutputTarget) -> Result<()>

Called to produce to a specific output target. This may be called without should_render if we are finalizing or clearing. This should flush if possible. Default implementation sends all to main output for backwards compatibility

fn aux_stream_is_tty(&self) -> bool

Check if auxillary stream is tty

fn terminal_size(&self) -> Result<Dimensions>

How big is the terminal to write to.