superconsole

Struct SuperConsole

pub struct SuperConsole { /* private fields */ }

Handles rendering the console using the user-defined Components and emitted messages. A Canvas area at the bottom of the terminal is re-rendered in place at each tick for the components, while a log area of emitted messages is produced above. Producing output from sources other than SuperConsole will break the TUI.

Implementations

impl SuperConsole

pub fn new() -> Option<Self>

Build a new SuperConsole with a root component.

pub fn forced_new(fallback_size: Dimensions) -> Self

Force a new SuperConsole to be built with a root component, regardless of whether the tty is compatible

pub fn compatible() -> bool

pub fn render(&mut self, root: &dyn Component) -> Result<()>

Render at a given tick. Draws all components and drains the emitted events buffer. This will produce any pending emitting events above the Canvas and will re-render the drawing area.

pub fn finalize(self, root: &dyn Component) -> Result<()>

Perform a final render with DrawMode::Final. Each component will have a chance to finalize themselves before the terminal is disposed of.

pub fn finalize_with_mode( self, root: &dyn Component, mode: DrawMode, ) -> Result<()>

Perform a final render, using a specified DrawMode. Each component will have a chance to finalize themselves before the terminal is disposed of.

pub fn emit_now(&mut self, lines: Lines, root: &dyn Component) -> Result<()>

Convenience method:

Calls queue_emit to add the lines.

Next, re-renders the superconsole.

Because this re-renders the console, it requires passed state. Overuse of this method can cause superconsole to use significant CPU.

pub fn emit(&mut self, lines: Lines)

Queues the passed lines to be drawn on the next render. The lines will not appear until the next render is called.

pub fn emit_aux(&mut self, lines: Lines)

Queues the passed lines of auxillary output to be drawn on the next render. The lines will not appear until the next render is called.

pub fn clear(&mut self) -> Result<()>

Clears the canvas portion of the superconsole.