superconsole::components

Module splitting

Splitting is one of the most primitive building blocks of a fully featured UI. This module contains components and enums that allow for splits along either dimension.

Structs

Split

Splits along a given direction for its child components. Child components are truncated to the bounds passed to them.

Enums

SplitKind

Controls the way the splitter displays its children.